refactor: extract the generation engine into generation.py (step 3b)
Move the 17-function generation engine (~2,440 lines) out of generate.py into
generation.py: the worker loop, question / user-continuation / answer
generation, the slop-fixer and anti-slop LLM calls, write_conversation,
system-prompt cycling, the API circuit-breaker / budget / requeue helpers, and
save_generation_state. generate.py drops from 4,939 to 2,564 lines (the
original monolith was 6,802 -> -62% overall).

The functions already referenced shared state via app_state (steps 2/3a), so
they move nearly verbatim. generation.py imports app_state + the existing
helper modules (api_handler / detection / text_utils / logging_config) and
NEVER imports generate.py. The single call back toward the GUI
(update_live_prompt_preview) is routed through
app_state.live_prompt_preview_hook via a small engine-side shim; generate.py
registers its GUI implementation as the hook. So the dependency is strictly
one-way: generate.py imports generation.

generate.py keeps the four entrypoints the lifecycle/GUI code calls via
`from generation import worker, check_budget_limit, estimate_time_remaining,
save_generation_state` (no call-site rewrites). The pre-existing
`undefined name 'text_context'` lives in a moved function and travels with it
(unchanged).

Verified: py_compile on all three modules; pyflakes shows generate.py with no
undefined names and generation.py with only the pre-existing text_context; and
a headless xvfb run confirming the one-way wiring, that the live-preview shim
routes through the registered hook, that engine functions execute post-move,
and that ConfigEditor still constructs.

NOTE: verified structurally + by import/construction; a live generation run
(worker threads + real API calls) was not exercised here and is worth a quick
smoke test on merge.
